requirements.txt
numpy
matplotlib.pyplot
sklearn.ensemble
sklearn.preprocessing 
collections
torch
torch.nn
torch.optim